@@ -1,49 +1,78 @@
 {
-    "name": "happy-dogs-proto",
-    "version": "0.0.0",
-    "scripts": {
-        "ng": "ng",
-        "start": "ng serve",
-        "build": "ng build",
-        "test": "jest",
-        "test:watch": "jest --watch",
-        "test:coverage": "jest --watch:coverage",
-        "lint": "ng lint",
-        "e2e": "ng e2e"
-    },
-    "private": true,
-    "dependencies": {
-        "@angular/animations": "~10.0.4",
-        "@angular/common": "~10.0.4",
-        "@angular/compiler": "~10.0.4",
-        "@angular/core": "~10.0.4",
-        "@angular/forms": "~10.0.4",
-        "@angular/platform-browser": "~10.0.4",
-        "@angular/platform-browser-dynamic": "~10.0.4",
-        "@angular/router": "~10.0.4",
-        "leaflet": "^1.6.0",
-        "ng-bullet": "^1.0.3",
-        "rxjs": "~6.5.5",
-        "ts-jest": "^26.3.0",
-        "tslib": "^2.0.0",
-        "zone.js": "~0.10.3"
-    },
-    "devDependencies": {
-        "@angular-devkit/build-angular": "~0.1000.3",
-        "@angular/cli": "~10.0.3",
-        "@angular/compiler-cli": "~10.0.4",
-        "@types/jasmine": "~3.5.0",
-        "@types/jasminewd2": "~2.0.3",
-        "@types/jest": "^26.0.10",
-        "@types/node": "^12.11.1",
-        "codelyzer": "^6.0.0",
-        "jasmine-core": "~3.5.0",
-        "jasmine-spec-reporter": "~5.0.0",
-        "jest": "^26.4.2",
-        "jest-preset-angular": "^8.2.1",
-        "protractor": "~7.0.0",
-        "ts-node": "~8.3.0",
-        "tslint": "~6.1.0",
-        "typescript": "~3.9.5"
+  "name": "happy-dogs-proto",
+  "version": "0.0.0",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve --proxy-config proxy.conf.json",
+    "build": "ng build",
+    "test": "jest",
+    "test:watch": "jest --watch",
+    "test:coverage": "jest --coverage",
+    "pretty": "pretty-quick",
+    "lint": "ng lint",
+    "e2e": "ng e2e"
+  },
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "~11.0.5",
+    "@angular/common": "~11.0.5",
+    "@angular/compiler": "~11.0.5",
+    "@angular/core": "~11.0.5",
+    "@angular/forms": "~11.0.5",
+    "@angular/platform-browser": "~11.0.5",
+    "@angular/platform-browser-dynamic": "~11.0.5",
+    "@angular/router": "~11.0.5",
+    "@clr/angular": "^4.0.8",
+    "@clr/core": "^4.0.8",
+    "@clr/icons": "^4.0.8",
+    "@clr/ui": "^4.0.8",
+    "@webcomponents/webcomponentsjs": "^2.5.0",
+    "leaflet": "^1.6.0",
+    "ng-bullet": "^1.0.3",
+    "rxjs": "~6.6.3",
+    "ts-jest": "^26.3.0",
+    "tslib": "^2.0.3",
+    "zone.js": "~0.10.3"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "^0.1100.5",
+    "@angular-eslint/builder": "1.1.0",
+    "@angular-eslint/eslint-plugin": "1.1.0",
+    "@angular-eslint/eslint-plugin-template": "1.1.0",
+    "@angular-eslint/schematics": "1.1.0",
+    "@angular-eslint/template-parser": "1.1.0",
+    "@angular/cli": "~11.0.5",
+    "@angular/compiler-cli": "~11.0.5",
+    "@types/jasmine": "^3.6.2",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/jest": "^26.0.10",
+    "@types/node": "^12.11.1",
+    "@typescript-eslint/eslint-plugin": "4.3.0",
+    "@typescript-eslint/parser": "4.3.0",
+    "codelyzer": "^6.0.0",
+    "eslint": "^7.6.0",
+    "eslint-plugin-import": "2.22.1",
+    "eslint-plugin-jsdoc": "30.7.6",
+    "eslint-plugin-prefer-arrow": "1.2.2",
+    "eslint-plugin-react": "latest",
+    "eslint-plugin-sonarjs": "^0.5.0",
+    "husky": "^4.3.6",
+    "jasmine-core": "~3.5.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "jest": "^26.6.3",
+    "jest-preset-angular": "^8.3.2",
+    "prettier": "^2.2.1",
+    "pretty-quick": "^3.1.0",
+    "protractor": "~7.0.0",
+    "ts-node": "~8.3.0",
+    "tslint": "~6.1.0",
+    "tslint-config-prettier": "^1.18.0",
+    "typescript": "~4.0.5"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "npm run pretty && git add .",
+      "pre-push": "npm run test && npm run build"
     }
+  }
 }
