@@ -1,6 +1,7 @@
-import {Component, OnInit, AfterViewInit} from '@angular/core';
-import {ApiService} from '@shared/services/api.service';
+import { Component, OnInit, AfterViewInit } from '@angular/core';
+import { ApiService } from '@shared/services/api.service';
 import * as L from 'leaflet';
+import { MarkerService } from '@shared/services/marker.service';
 
 @Component({
   selector: 'app-map',
@@ -10,18 +11,16 @@ import * as L from 'leaflet';
 export class MapComponent implements OnInit, AfterViewInit {
   map: L.map;
 
-  constructor(private apiS: ApiService) {
-  }
+  constructor(private markerService: MarkerService) {}
 
-  ngOnInit(): void {
-  }
+  ngOnInit(): void {}
 
   ngAfterViewInit(): void {
     this.initMap();
     this.populateMap();
     this.getGeoLocation();
-    this.apiS.addWaterFountains(this.map);
-    this.apiS.addDogzones(this.map);
+    this.markerService.addWaterFountains(this.map);
+    this.markerService.addDogzones(this.map);
     // this.apiS.addPoobags(this.map);
   }
 
@@ -30,26 +29,23 @@ export class MapComponent implements OnInit, AfterViewInit {
   }
 
   private populateMap(): void {
-    const tiles = L.tileLayer(
-      'https://api.mapbox.com/styles/v1/{id}/tiles/{z}/{x}/{y}?access_token={accessToken}',
-      {
-        maxZoom: 25,
-        id: 'mapbox/streets-v11',
-        tileSize: 256,
-        attribution:
-          '&copy; <a href="http://www.openstreetmap.org/copyright">OpenStreetMap</a>',
-        accessToken: 'pk.eyJ1IjoiZG9taW5pa2RvcmZzdGV0dGVyIiwiYSI6ImNrZTQyOWk1NTBvankycW1zbHB3ZmI1ZzAifQ.Nn6ldXhhXF50Y2WNzdVgTg'
-      }
-    );
+    const tiles = L.tileLayer('https://api.mapbox.com/styles/v1/{id}/tiles/{z}/{x}/{y}?access_token={accessToken}', {
+      maxZoom: 25,
+      id: 'mapbox/streets-v11',
+      tileSize: 256,
+      attribution: '&copy; <a href="http://www.openstreetmap.org/copyright">OpenStreetMap</a>',
+      accessToken:
+        'pk.eyJ1IjoiZG9taW5pa2RvcmZzdGV0dGVyIiwiYSI6ImNrZTQyOWk1NTBvankycW1zbHB3ZmI1ZzAifQ.Nn6ldXhhXF50Y2WNzdVgTg',
+    });
 
     tiles.addTo(this.map);
   }
 
   private getGeoLocation(): void {
     const map = this.map;
-    map.locate({setView: true});
+    map.locate({ setView: true });
 
-    // tslint:disable-next-line:typedef
+    // eslint-disable-next-line
     function onLocationFound(e) {
       const radius = e.accuracy / 2;
       const circle = L.circle(e.latlng, radius);
@@ -61,5 +57,4 @@ export class MapComponent implements OnInit, AfterViewInit {
 
     map.on('locationfound', onLocationFound);
   }
-
 }
