@@ -1,6 +1,6 @@
 import * as L from 'leaflet';
 
-const IconTemplate = L.Icon.extend({
+const iconTemplate = L.Icon.extend({
   options: {
     iconSize: [30, 80],
     iconAnchor: [22, 94],
@@ -9,12 +9,12 @@ const IconTemplate = L.Icon.extend({
   },
 });
 
-export const ICON_DOGGYBAG = new IconTemplate({
+export const ICON_DOGGYBAG = new iconTemplate({
   iconUrl: '/assets/images/marker/poobag_pin.svg',
 });
-export const ICON_DOGZONE = new IconTemplate({
+export const ICON_DOGZONE = new iconTemplate({
   iconUrl: '/assets/images/marker/dogzone_pin.svg',
 });
-export const ICON_WATERFOUNTAIN = new IconTemplate({
+export const ICON_WATERFOUNTAIN = new iconTemplate({
   iconUrl: '/assets/images/marker/water_pin.svg',
 });
