@@ -1,16 +1,17 @@
-import {ApiService} from '@shared/services/api.service';
-import {MarkerService} from '@shared/services/marker.service';
-import {HttpClient} from '@angular/common/http';
-import {of} from 'rxjs';
-import {APIResponse} from '@shared/models/api-response.model';
+import { ApiService } from '@shared/services/api.service';
+import { MarkerService } from '@shared/services/marker.service';
+import { HttpClient } from '@angular/common/http';
+import { of } from 'rxjs';
+import { MapMarkerResponse } from '@shared/models/map-marker.model';
 import * as L from 'leaflet';
-import {TestBed} from '@angular/core/testing';
-import {LocalStorageMock} from '@shared/mock/localstorage.mock';
+import { TestBed } from '@angular/core/testing';
+import { LocalStorageMock } from '@shared/mock/localstorage.mock';
 
 const MOCK_FEATURE = {
   geometry: {
     coordinates: [0, 0],
   },
+  // eslint-disable-next-line @typescript-eslint/naming-convention
   geometry_name: 'test',
   id: 'test',
   properties: {},
@@ -25,13 +26,13 @@ const MOCK_RESPONSE = {
   features: [MOCK_FEATURE, MOCK_FEATURE],
   totalFeatures: 2,
   type: 'test',
-} as APIResponse;
+} as MapMarkerResponse;
 
 describe('ApiService', () => {
   let service: ApiService;
   let markerS: MarkerService;
   let http: HttpClient;
-  // tslint:disable-next-line:prefer-const
+  // eslint-disable-next-line prefer-const
   let map: L.map;
   let httpGetSpy: jest.SpyInstance;
   let addMarkerToMapSpy: jest.SpyInstance;
@@ -41,25 +42,20 @@ describe('ApiService', () => {
       get: () => jest.fn(),
     } as any;
     markerS = {
-      addMarkerToMap: () => {
-      },
+      addMarkerToMap: () => {},
     } as any;
 
     httpGetSpy = jest.spyOn(http, 'get');
     addMarkerToMapSpy = jest.spyOn(markerS, 'addMarkerToMap');
     httpGetSpy.mockReturnValue(of(MOCK_RESPONSE));
-    addMarkerToMapSpy.mockImplementation(() => {
-    });
+    addMarkerToMapSpy.mockImplementation(() => {});
 
-    Object.defineProperty(global, 'localStorage', {writable: true});
+    Object.defineProperty(global, 'localStorage', { writable: true });
     // noinspection JSConstantReassignment
     global.localStorage = new LocalStorageMock();
 
     TestBed.configureTestingModule({
-      providers: [
-        ApiService,
-        {provide: MarkerService, useValue: markerS},
-        {provide: HttpClient, useValue: http}]
+      providers: [ApiService, { provide: MarkerService, useValue: markerS }, { provide: HttpClient, useValue: http }],
     });
     service = TestBed.inject(ApiService);
   });
@@ -75,6 +71,7 @@ describe('ApiService', () => {
         {
           ...MOCK_FEATURE,
           properties: {
+            // eslint-disable-next-line @typescript-eslint/naming-convention
             NAME: 'Auslaufbrunnen',
           },
         },
@@ -82,18 +79,14 @@ describe('ApiService', () => {
     };
 
     it('should call http get once, without cache', () => {
-      httpGetSpy.mockReturnValueOnce(
-        of(API_RESPONSE)
-      );
-      service.addWaterFountains(map);
+      httpGetSpy.mockReturnValueOnce(of(API_RESPONSE));
+      markerS.addWaterFountains(map);
       expect(httpGetSpy).toBeCalledTimes(1);
     });
 
     it('should add 2 markers to the map, without cache', () => {
-      httpGetSpy.mockReturnValueOnce(
-        of(API_RESPONSE)
-      );
-      service.addWaterFountains(map);
+      httpGetSpy.mockReturnValueOnce(of(API_RESPONSE));
+      markerS.addWaterFountains(map);
       expect(addMarkerToMapSpy).toBeCalledTimes(1);
     });
 
@@ -102,19 +95,17 @@ describe('ApiService', () => {
       // first set local storage
       localStorage.setItem(storageKey, JSON.stringify(API_RESPONSE));
       // set http get response also
-      httpGetSpy.mockReturnValueOnce(
-        of(API_RESPONSE)
-      );
+      httpGetSpy.mockReturnValueOnce(of(API_RESPONSE));
       const lsGetItemSpy = jest.spyOn(localStorage, 'getItem');
 
-      service.addWaterFountains(map);
+      markerS.addWaterFountains(map);
       expect(lsGetItemSpy).toHaveBeenCalledWith(storageKey);
     });
   });
 
   describe('addPoobags', () => {
     beforeEach(() => {
-      service.addPoobags(map);
+      markerS.addPoobags(map);
     });
 
     it('should call http get once', () => {
@@ -133,6 +124,7 @@ describe('ApiService', () => {
         {
           ...MOCK_FEATURE,
           properties: {
+            // eslint-disable-next-line @typescript-eslint/naming-convention
             TYP: 'Hundezone',
           },
         },
@@ -140,10 +132,8 @@ describe('ApiService', () => {
     };
 
     beforeEach(() => {
-      httpGetSpy.mockReturnValueOnce(
-        of(API_RESPONSE)
-      );
-      service.addDogzones(map);
+      httpGetSpy.mockReturnValueOnce(of(API_RESPONSE));
+      markerS.addDogzones(map);
     });
 
     it('should call http get once', () => {
