@@ -1,4 +1,4 @@
-import {MarkerService} from '@shared/services/marker.service';
+import { MarkerService } from '@shared/services/marker.service';
 import * as L from 'leaflet';
 
 describe('MarkerService', () => {
@@ -6,7 +6,7 @@ describe('MarkerService', () => {
   let markerS: MarkerService;
 
   beforeEach(() => {
-    markerS = new MarkerService();
+    markerS = new MarkerService(mapMock);
     mapMock = {};
   });
 
@@ -16,12 +16,10 @@ describe('MarkerService', () => {
 
   it('should add marker to map', () => {
     mapMock = {
-      addLayer: () => {
-      }
+      addLayer: () => {},
     } as any;
     const addMarkerToMapSpy = jest.spyOn(markerS, 'addMarkerToMap');
     markerS.addMarkerToMap(mapMock, [1, 1], {} as L.Icon);
     expect(addMarkerToMapSpy).toHaveBeenCalled();
   });
-
 });
