@@ -1,23 +1,19 @@
 import 'jest-preset-angular';
 
-Object.defineProperty(window, 'CSS', {value: null});
+Object.defineProperty(window, 'CSS', { value: null });
 Object.defineProperty(window, 'getComputedStyle', {
-  value: () => {
-    return {
-      display: 'none',
-      appearance: ['-webkit-appearance']
-    };
-  }
+  value: () => ({
+    display: 'none',
+    appearance: ['-webkit-appearance'],
+  }),
 });
 
 Object.defineProperty(document, 'doctype', {
-  value: '<!DOCTYPE html>'
+  value: '<!DOCTYPE html>',
 });
 Object.defineProperty(document.body.style, 'transform', {
-  value: () => {
-    return {
-      enumerable: true,
-      configurable: true
-    };
-  }
+  value: () => ({
+    enumerable: true,
+    configurable: true,
+  }),
 });
